Move SystemImageType to constants module
