+++ datastore/src/test/java/org/jboss/pnc/datastore/repositories/BuildConfigurationRepositoryImplTest.java
@@ -29,6 +29,7 @@ import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
 import javax.inject.Inject;
+
 import java.util.ArrayList;
 import java.util.List;
