+++ integration-test/src/test/java/org/jboss/pnc/integration/EnvironmentRestTest.java
@@ -31,7 +31,7 @@ import org.jboss.pnc.integration.client.EnvironmentRestClient;
 import org.jboss.pnc.integration.client.util.RestResponse;
 import org.jboss.pnc.integration.deployments.Deployments;
 import org.jboss.pnc.integration.utils.ResponseUtils;
-import org.jboss.pnc.model.SystemImageType;
+import org.jboss.pnc.enums.SystemImageType;
 import org.jboss.pnc.rest.endpoint.BuildEnvironmentEndpoint;
 import org.jboss.pnc.rest.provider.BuildEnvironmentProvider;
 import org.jboss.pnc.rest.restmodel.BuildEnvironmentRest;
@@ -49,6 +49,7 @@ import java.lang.invoke.MethodHandles;
 import java.util.List;
 
 import static com.jayway.restassured.RestAssured.given;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.jboss.pnc.integration.env.IntegrationTestEnv.getHttpPort;
 import static org.jboss.pnc.test.util.JsonUtils.toJson;
