+++ model/src/test/java/org/jboss/pnc/model/BuildConfigurationDependencyTest.java
@@ -17,6 +17,8 @@
  */
 package org.jboss.pnc.model;
 
+import org.jboss.pnc.enums.SystemImageType;
+
 import java.util.List;
 
 import javax.persistence.EntityManager;
