+++ pnc-mock/src/main/java/org/jboss/pnc/mock/environmentdriver/EnvironmentDriverMock.java
@@ -17,7 +17,7 @@
  */
 package org.jboss.pnc.mock.environmentdriver;
 
-import org.jboss.pnc.model.SystemImageType;
+import org.jboss.pnc.enums.SystemImageType;
 import org.jboss.pnc.spi.builddriver.DebugData;
 import org.jboss.pnc.spi.environment.EnvironmentDriver;
 import org.jboss.pnc.spi.environment.RunningEnvironment;
@@ -26,6 +26,7 @@ import org.jboss.pnc.spi.environment.exception.EnvironmentDriverException;
 import org.jboss.pnc.spi.repositorymanager.model.RepositorySession;
 
 import javax.enterprise.context.ApplicationScoped;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
