+++ pnc-mock/src/main/java/org/jboss/pnc/mock/environmentdriver/EnvironmentDriverWithFailedContainerInitializationMock.java
@@ -18,7 +18,7 @@
 
 package org.jboss.pnc.mock.environmentdriver;
 
-import org.jboss.pnc.model.SystemImageType;
+import org.jboss.pnc.enums.SystemImageType;
 import org.jboss.pnc.spi.builddriver.DebugData;
 import org.jboss.pnc.spi.environment.EnvironmentDriver;
 import org.jboss.pnc.spi.environment.RunningEnvironment;
@@ -27,6 +27,7 @@ import org.jboss.pnc.spi.environment.exception.EnvironmentDriverException;
 import org.jboss.pnc.spi.repositorymanager.model.RepositorySession;
 
 import javax.enterprise.context.ApplicationScoped;
+
 import java.util.function.Consumer;
 
 /**
