+++ spi/src/main/java/org/jboss/pnc/spi/environment/EnvironmentDriver.java
@@ -17,7 +17,7 @@
  */
 package org.jboss.pnc.spi.environment;
 
-import org.jboss.pnc.model.SystemImageType;
+import org.jboss.pnc.enums.SystemImageType;
 import org.jboss.pnc.spi.builddriver.DebugData;
 import org.jboss.pnc.spi.environment.exception.EnvironmentDriverException;
 import org.jboss.pnc.spi.repositorymanager.model.RepositorySession;
