+++ spi/src/main/java/org/jboss/pnc/spi/executor/BuildExecutionConfiguration.java
@@ -19,7 +19,7 @@
 package org.jboss.pnc.spi.executor;
 
 import org.jboss.pnc.enums.BuildType;
-import org.jboss.pnc.model.SystemImageType;
+import org.jboss.pnc.enums.SystemImageType;
 import org.jboss.pnc.spi.repositorymanager.ArtifactRepository;
 import org.jboss.pnc.spi.repositorymanager.BuildExecution;
